reorganize common netty client code
